# Runner: fix solver's random seed