@jacek-oet 2024-10-21
> Run the benchmark runner on two small/quick benchmarks with iterations=10, both before and after the changes.
> If everything works as expected, we should see the standard deviation of runtime decrease due to the introduction of the seed.


The default seed for SCIP, HiGHS, and Gurobi is set to 0
For GLPK, the seed is not automatically set to 0 unless specified.


**Scip**

```
# global shift of all random seeds in the plugins and the LP random seed
# [type: int, advanced: FALSE, range: [0,2147483647], default: 0]
randomization/randomseedshift = 0
```

ref: https://www.scipopt.org/doc/html/PARAMETERS.php

**Highs**

```
[random_seed](https://ergo-code.github.io/HiGHS/dev/options/definitions/#random_seed)
Random seed used in HiGHS
Type: integer
Range: {0, 2147483647}
Default: 0
```
ref: https://ergo-code.github.io/HiGHS/dev/options/definitions/#random_seed

**Gurobi**
```
Seed Random number seed  	
Type:	int  	
Default value:	0  	
Minimum value:	0  	
Maximum value:	MAXINT
```
ref: https://www.gurobi.com/documentation/current/refman/seed.html

**Glpk**

```
 --seed value      initialize pseudo-random number generator used in
                    MathProg model with specified seed (any integer);
                    if seed value is ?, some random seed will be used
```

ref: https://en.wikibooks.org/wiki/GLPK/Using_GLPSOL


here is the test for **glpk**

config
```
       {
            "name": "problem_5.lp",
            "url": "https://drive.usercontent.google.com/download?id=1D0_mo--5r9m46F05hjHpdzGDoV0fbsfd&export=download&authuser=0",
        },
        {
            "name": "problem_6.lp",
            "url": "https://drive.usercontent.google.com/download?id=1D0_mo--5r9m46F05hjHpdzGDoV0fbsfd&export=download&authuser=0",
        },
```

when we unset seed
```
Benchmark,Solver,Status,Termination Condition,Objective Value,Runtime Mean (s),Runtime StdDev (s),Memory Mean (MB),Memory StdDev (MB)
problem_5,glpk,ok,optimal,84667526620.0,174.73571794033052,0.9553934845566686,273.2372,0.7318823220527491
problem_6,glpk,ok,optimal,84667526620.0,174.8346357345581,0.7375039599045516,273.7084,0.3594190373984685
```

when seed = 0
```
Benchmark,Solver,Status,Termination Condition,Objective Value,Runtime Mean (s),Runtime StdDev (s),Memory Mean (MB),Memory StdDev (MB)
problem_5,glpk,ok,optimal,84667526620.0,174.27955131530763,0.5133412684618199,272.152,0.4724479983142171
problem_6,glpk,ok,optimal,84667526620.0,173.8979598045349,0.4519245373436955,272.1988,0.40169883688597957
```

The standard deviation of runtime decreases when the seed is introduced (seed = 0).

some additional test for seed = 0 and seed = 2147483647
seed = 0:

```
Benchmark,Solver,Status,Termination Condition,Objective Value,Runtime Mean (s),Runtime StdDev (s),Memory Mean (MB),Memory StdDev (MB)
problem_5,highs,ok,optimal,84667526618.30806,81.57768340110779,0.31467671954550286,439.4376,0.9171282716537926
problem_5,glpk,ok,optimal,84667526620.0,174.27955131530763,0.5133412684618199,272.152,0.4724479983142171
problem_6,scip,ok,optimal,84667526618.308,88.37553451061248,0.1368302587144211,649.0548,0.14281051937600422
problem_6,highs,ok,optimal,84667526618.30806,81.39963517189025,0.10264529897350569,439.0476,1.3959604578926963
problem_6,glpk,ok,optimal,84667526620.0,173.8979598045349,0.4519245373436955,272.1988,0.40169883688597957
problem_6,scip,ok,optimal,84667526618.308,88.07555766105652,0.181675083289328,643.5348,17.949819341709258
```

seed = 2147483647

```
Benchmark,Solver,Status,Termination Condition,Objective Value,Runtime Mean (s),Runtime StdDev (s),Memory Mean (MB),Memory StdDev (MB)
problem_5,highs,ok,optimal,84667526618.30806,81.78446354866028,0.2820689765107425,418.85360000000003,12.258789818828866
problem_5,glpk,ok,optimal,84667526620.0,173.3591020822525,0.26864107658940656,272.9236,0.678923199714887
problem_5,scip,ok,optimal,84667526618.308,87.91752769947053,0.18133391029243084,643.9432,18.728743712984762
problem_6,highs,ok,optimal,84667526618.30806,81.61518692970276,0.15161383059101768,423.144,1.1407989403142866
problem_6,glpk,ok,optimal,84667526620.0,173.69413158893585,0.5108453023436176,273.3404,0.49529860802460307
problem_6,scip,ok,optimal,84667526618.308,87.7999925851822,0.19708740832486268,650.108,0.19377879714308102
```