@jacek-oet 2024-10-22
@siddharth-krishna Here is the result after running

**Config**:
- problem_5
https://drive.usercontent.google.com/download?id=1SrFi3qDK6JpUM-pzyyz11c8PzFq74XEO&export=download&authuser=0

- problem_6
https://drive.usercontent.google.com/download?id=1SrFi3qDK6JpUM-pzyyz11c8PzFq74XEO&export=download&authuser=0



**Result**:

- seed = 0 

```
Benchmark,Solver,Status,Termination Condition,Objective Value,Runtime Mean (s),Runtime StdDev (s),Memory Mean (MB),Memory StdDev (MB)
problem_6,glpk,ok,optimal,84667526620.0,174.0086378335953,0.9573490493835888,273.6616,0.6501314738830497
problem_6,highs,ok,optimal,84667526618.30806,81.3053231716156,0.12376560562032854,441.0248,0.6285837343672859
problem_6,scip,ok,optimal,84667526618.308,88.28385570049286,0.2673504972502768,650.6584,0.15564289611514281
problem_5,glpk,warning,unknown,nan,0.3523340463638306,0.0024787299265993335,167.208,0.16091957963875023
problem_5,highs,ok,optimal,31714138111.861046,7.370548486709595,0.026833936385292686,444.0536,0.8803094405441245
problem_5,scip,ok,optimal,31714138111.860943,104.89062049388886,0.25440500990600684,736.7088,0.33271836605621286
```

- seed = random.randint(0,100) 
```
Benchmark,Solver,Status,Termination Condition,Objective Value,Runtime Mean (s),Runtime StdDev (s),Memory Mean (MB),Memory StdDev (MB)
problem_6,glpk,ok,optimal,84667526620.0,173.62057974338532,0.5281723069030282,266.5592,20.005901067212918
problem_6,highs,ok,optimal,84667526618.30806,81.96060452461242,1.1333293305456815,431.14160000000004,10.295199865309403
problem_6,scip,ok,optimal,84667526618.308,101.0902177810669,21.130594996272105,641.082,19.741337948578877
problem_5,glpk,warning,unknown,nan,0.35806984901428224,0.010311103614394418,166.7908,0.12536240797517348
problem_5,highs,ok,optimal,31714138111.861046,8.53553512096405,0.937450164895183,457.3424,19.299454564543762
problem_5,scip,ok,optimal,31714138111.860943,117.2516789674759,11.766968568854994,737.318,3.6230144600564045
```

**Summary**
- Seed = 0
The standard deviation (StdDev) of runtime is quite low across all solvers (GLPK, HiGHS, SCIP) for both problem_6 and problem_5. This indicates that the solver’s performance is consistent when a fixed seed is used.
Memory usage also remains consistent with low variability (small StdDev) when the seed is fixed.
Predictable and consistent results, with minimal variation in both runtime and memory usage. This shows that using a fixed seed makes the solver performance deterministic, which is useful for benchmarking and comparisons.

- Seed = random.randint(0, 100)
There is an increase in the standard deviation of runtime and memory compared to the fixed seed case.
Introducing randomness through seed variations increases both runtime and memory variability, particularly for SCIP and GLPK. This demonstrates the solvers' sensitivity to the initial conditions set by the seed, with potential impacts on solver efficiency.